pyproject.toml
[tool.poetry]
name = "christinanissi"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = ">=3.8,<4"
django = ">=2.2"
djangoql = "*"
django-extensions = "*"
django-redis = "*"
psycopg2-binary = "*"
pytest = "*"
pytest-cov = "*"
sentry-sdk = "*"
shortuuid = "*"
unittest-xml-reporting = "*"
uwsgi = "*"
werkzeug = "*"
whitenoise = "*"
urllib3 = "^1.26.12"
django-environ = "^0.9.0"
django-taggit = "^3.0.0"
django-summernote = "^0.8.20.0"

[tool.poetry.dev-dependencies]
ipython = "*"
bpython = "*"
pudb = "*"
django-stubs = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
